% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_check.R
\name{autotest}
\alias{autotest}
\alias{spell_check_setup}
\alias{spell_check_test}
\title{Auto Spell Check}
\usage{
spell_check_setup(path = ".", vignettes = TRUE, lang = "en_US")
}
\arguments{
\item{path}{path to file or package root directory containing the \code{DESCRIPTION} file}

\item{vignettes}{spell check \code{rmd} and \code{rnw} files in the \code{vignettes} folder}

\item{lang}{string with dictionary language string for \link[hunspell:dictionary]{hunspell::dictionary}}
}
\description{
Setup automatic spell checking as a unit test for \code{CMD check}.
}
\details{
Use \code{\link[=spell_check_setup]{spell_check_setup()}} to add a unit test to your package which automatically
spell checks documentation and vignettes when running \code{R CMD check}.
}
\seealso{
Other spelling: \code{\link{spell_check}},
  \code{\link{wordlist}}
}
