% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell-check-project.R
\name{change_language}
\alias{change_language}
\title{Change the language for spell checking for a subset of files in a project.}
\usage{
change_language(root, path, lang)
}
\arguments{
\item{root}{path to the project root.}

\item{path}{the relative path to a file or directory within the project}

\item{lang}{the language to use when spell checking the file or files with
the directory.
Use \code{lang = "ignore"} to exclude \code{path} from spell checking.}
}
\description{
Change the language for spell checking for a subset of files in a project.
}
