% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell-check-bookdown.R
\name{spell_check_bookdown}
\alias{spell_check_bookdown}
\alias{spelling}
\title{Package Spell Checking}
\usage{
spell_check_bookdown(path = ".", lang = NULL, use_wordlist = TRUE)
}
\arguments{
\item{path}{path to package root directory containing the \code{DESCRIPTION} file}

\item{lang}{set \code{Language} field in \code{DESCRIPTION} e.g. \code{"en-US"} or \code{"en-GB"}.
For supporting other languages, see the \href{https://bit.ly/2EquLKy}{hunspell vignette}.}

\item{use_wordlist}{ignore words in the package \link[=get_wordlist]{WORDLIST} file}
}
\description{
Automatically spell-check bookdown source files.
}
\details{
Parses and checks bookdown source files, and if these files are present
\itemize{
\item README.md (when there's no README.Rmd)
\item NEWS.md
\item text fields in the \code{DESCRIPTION} file if present.
}

The preferred spelling language (typically \code{en-GB} or \code{en-US}) should be
\itemize{
\item specified in the \code{Language} field from your package \code{DESCRIPTION},
\item or entered as the \code{lang} argument.
To whitelist custom words
use the bookdown \link[=get_wordlist]{WORDLIST} file which will be added to the dictionary
when spell checking. See \link{update_wordlist} to automatically populate and update this
file.
}

Hunspell includes dictionaries for \code{en_US} and \code{en_GB} by default. Other languages
require installation of a custom dictionary, see \link[hunspell:hunspell]{hunspell} for details.
}
\seealso{
Other spelling: \code{\link{spell_check_files}},
  \code{\link{spell_check_package}}, \code{\link{wordlist}}
}
\concept{spelling}
