% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell-check-project.R
\name{spell_check_project}
\alias{spell_check_project}
\title{Spell checking of a project with markdown files}
\usage{
spell_check_project(root = ".", lang)
}
\arguments{
\item{root}{path to the project root.}

\item{lang}{default language to use for spell checking.
When omitted it refers to language used during the last time the project was
spell checked.
The fall-back language is \code{en-US}.}
}
\description{
Detect all Rmarkdown and markdown files recursively from the project root.
It ignores files in the \code{renv} folder as they most likely contain vignettes
from archived packages.
}
