% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlist.R
\name{wordlist}
\alias{wordlist}
\alias{update_wordlist}
\alias{get_wordlist}
\title{The WORDLIST file}
\usage{
update_wordlist(pkg = ".", vignettes = TRUE, lang = "en_US",
  confirm = TRUE)

get_wordlist(pkg = ".")
}
\arguments{
\item{pkg}{path to package root directory containing the \code{DESCRIPTION} file}

\item{vignettes}{spell check \code{rmd} and \code{rnw} files in the \code{vignettes} folder}

\item{lang}{string with dictionary language string for \link[hunspell:dictionary]{hunspell::dictionary}}

\item{confirm}{show changes and ask confirmation before adding new words to the list}
}
\description{
Read and update the wordlist included with a package. The wordlist is stored in
the file \code{inst/WORDLIST} in the source package and is used to allow custom words
which will be added to the dictionary when spell checking.
}
\seealso{
Other spelling: \code{\link{spell_check_files}},
  \code{\link{spell_check_package}}
}
