% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_check.R
\name{spell_test_run}
\alias{spell_test_run}
\alias{spell_test_setup}
\title{Auto Spell Check}
\usage{
spell_test_run(vignettes = TRUE, lang = "en_US")

spell_test_setup(path = ".", vignettes = TRUE, lang = "en_US")
}
\arguments{
\item{vignettes}{spell check \code{rmd} and \code{rnw} files in the \code{vignettes} folder}

\item{lang}{string with dictionary language string for \link[hunspell:dictionary]{hunspell::dictionary}}

\item{path}{path to file or package root directory containing the \code{DESCRIPTION} file}
}
\description{
Setup automatic spell checking as a unit test for \code{CMD check}.
}
\details{
The \code{\link[=spell_test_run]{spell_test_run()}} function is intended to be called as a unit test triggered
via running \code{CMD check} on a package. Use \code{\link[=spell_test_setup]{spell_test_setup()}} to automatically
create the unit test that does this.
}
\seealso{
Other spelling: \code{\link{spell_check}},
  \code{\link{wordlist}}
}
