% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-markdown.R
\name{parse_text}
\alias{parse_text}
\alias{parse_text_md}
\title{Text Parsers}
\usage{
parse_text_md(path, extensions = TRUE, yaml_fields = c("title",
  "subtitle", "description"))
}
\arguments{
\item{path}{markdown file}

\item{extensions}{render markdown extensions? Passed to \link[commonmark:markdown_xml]{commonmark}}

\item{yaml_fields}{character vector indicating which fields of the yaml
front matter should be spell checked.}
}
\description{
Parse text from various formats and return a data frame with text lines
and position in the source document.
}
