% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell-check.R
\name{spell_check_package}
\alias{spell_check_package}
\title{Spell Check Package}
\usage{
spell_check_package(pkg = ".", ignore = character(), dict = "en_US")
}
\arguments{
\item{pkg}{package description, can be path or package name. Passed to \link[devtools:as.package]{devtools::as.package}.}

\item{ignore}{character vector with words to ignore. Passed to \link[hunspell:hunspell]{hunspell}.}

\item{dict}{a dictionary object or language string. Passed to \link[hunspell:hunspell]{hunspell}.}
}
\description{
Runs a spell check on text fields in the package description file and manual pages.
Hunspell includes dictionaries for \code{en_US} and \code{en_GB} by default. Other languages
require installation of a custom dictionary, see \link[hunspell:hunspell]{hunspell} for details.
}
