% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell-check.R
\name{spell_check}
\alias{spell_check}
\alias{spell_check_package}
\alias{spell_check_files}
\alias{spell_check_text}
\title{Spell Checking}
\usage{
spell_check_package(path = ".", vignettes = TRUE, lang = "en_US",
  use_wordlist = TRUE)

spell_check_files(path, ignore = character(), lang = "en_US")

spell_check_text(text, ignore = character(), lang = "en_US")
}
\arguments{
\item{path}{path to file or package root directory containing the \code{DESCRIPTION} file}

\item{vignettes}{spell check \code{rmd} and \code{rnw} files in the \code{vignettes} folder}

\item{lang}{string with dictionary language string for \link[hunspell:dictionary]{hunspell::dictionary}}

\item{use_wordlist}{ignore words in the package \link[=get_wordlist]{WORDLIST}}

\item{ignore}{character vector with words to ignore in \link[hunspell:hunspell]{hunspell}}

\item{text}{character vector with plain text}
}
\description{
Perform a spell check on manual pages, vignettes, description files and other formats.
}
\details{
The \link{spell_check_package} function checks the package manual pages, rmd/rnw vignettes,
and text fields in the \code{DESCRIPTION} file. The \link[=get_wordlist]{WORDLIST} can be used to whitelist
custom words in your package, which will be added to the dictionary when spell checking.

Hunspell includes dictionaries for \code{en_US} and \code{en_GB} by default. Other languages
require installation of a custom dictionary, see \link[hunspell:hunspell]{hunspell} for details.
}
\examples{
# Example files
files <- list.files(system.file("examples", package = "knitr"),
  pattern = "\\\\.(Rnw|Rmd|html)$", full.names = TRUE)
spell_check_files(files)
}
