% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell-check-project.R
\name{spell_check_project}
\alias{spell_check_project}
\title{Spell checking of a project with markdown files}
\usage{
spell_check_project(root = ".", lang = "en_US")
}
\arguments{
\item{root}{path to the project root.}

\item{lang}{set \code{Language} field in \code{DESCRIPTION} e.g. \code{"en-US"} or \code{"en-GB"}.}
}
\description{
Detect all Rmarkdown and markdown files recursively from the project root.
It ignores files in the \code{renv} folder as they most likely contain vignettes
from archived packages.
}
