% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlist.R
\name{wordlist}
\alias{wordlist}
\alias{update_wordlist}
\alias{get_wordlist}
\title{Update Wordlist}
\usage{
update_wordlist(path = ".", vignettes = TRUE, lang = "en_US",
  confirm = TRUE)

get_wordlist(path = ".")
}
\arguments{
\item{path}{path to file or package root directory containing the \code{DESCRIPTION} file}

\item{vignettes}{spell check \code{rmd} and \code{rnw} files in the \code{vignettes} folder}

\item{lang}{string with dictionary language string for \link[hunspell:dictionary]{hunspell::dictionary}}

\item{confirm}{show changes and ask confirmation before writing changes}
}
\description{
Read and update the wordlist included with a package. The wordlist is stored in
the file \code{inst/WORDLIST} in the source package and is be used to whitelist
custom words that will be added to the dictionary when spell checking.
}
